Added a filter param for efficient query
Added a filter param for efficient query to avoid massive number of results that triggers older data to be excluded in the resulting json. Newrelic has a maximum number of data rows of 200, and if it goes beyond the limit theres a great chance that older instances wont be included in the resulting json and you need to request again for page 2. Adding filter param for name instance would eliminate the issue because it searches for the specific name instance